+++ tile_engine/ops/gemm/gemm_universal/gemm_universal_benchmark.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python3
 # Copyright (c) Advanced Micro Devices, Inc., or its affiliates.
 # SPDX-License-Identifier: MIT
 
@@ -11,7 +12,6 @@ import time
 import importlib.util
 from pathlib import Path
 from typing import List, Dict, Tuple, Optional
-
 
 def _import_gemm_benchmark():
     """Import validation utilities from commons directory."""
@@ -46,14 +46,13 @@ def _import_benchmark_utils():
 GemmBenchmark = _import_gemm_benchmark()
 benchmark_utils = _import_benchmark_utils()
 
-class GemmPreshuffleBenchmark(GemmBenchmark):
+class GemmUniversalBenchmark(GemmBenchmark):
     def __init__(self, build_dir: str, verbose: bool = False):
-        super().__init__(build_dir, verbose, name="benchmark_gemm_preshuffle_")
+        super().__init__(build_dir, verbose, name="benchmark_gemm_")
+
 
 def main():
-    parser = argparse.ArgumentParser(
-        description="GEMM Preshuffle Kernel Benchmarking Tool"
-    )
+    parser = argparse.ArgumentParser(description="GEMM Kernel Benchmarking Tool")
     parser.add_argument(
         "build_dir", help="Build directory containing kernel executables"
     )
@@ -68,9 +67,7 @@ def main():
     )
     parser.add_argument("--verify", action="store_true", help="Enable verification")
     parser.add_argument(
-        "--csv",
-        default="gemm_preshuffle_benchmark_results.csv",
-        help="CSV output filename",
+        "--csv", default="gemm_benchmark_results.csv", help="CSV output filename"
     )
     parser.add_argument(
         "--best", default="best_kernels.txt", help="Best kernels output filename"
@@ -115,10 +112,10 @@ def main():
             return 1
 
     # Create benchmark instance
-    benchmark = GemmPreshuffleBenchmark(args.build_dir, verbose=args.verbose)
+    benchmark = GemmUniversalBenchmark(args.build_dir, verbose=args.verbose)
 
     # Run benchmark sweep
-    print("Starting GEMM Preshuffle kernel benchmark sweep...")
+    print("Starting GEMM kernel benchmark sweep...")
     start_time = time.time()
 
     best_kernels = benchmark.benchmark_sweep(
